requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.3
cramjam==2.8.4
cryptography==43.0.1
cssselect==1.2.0
cssutils==2.11.1
distlib==0.3.9
dnspython==1.16.0
email_validator==2.2.0
emails==0.6
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==3.0.1
more-itertools==10.5.0
moto==2.2.0
motor==3.6.0
multidict==6.1.0
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
premailer==3.10.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pykerberos==1.2.4
pymongo==4.10.1
pyproject-api==1.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-snappy==0.7.3
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
responses==0.25.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.39.2
tomli==2.0.2
tox==4.21.2
twilio==9.3.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
xmltodict==0.14.1
yarl==1.15.2
